Review comments fixes
