+++ spring-cloud-dataflow-server-core/src/main/resources/schema-postgresql.sql
@@ -14,7 +14,7 @@ CREATE TABLE DEPLOYMENT_IDS  (
 	DEPLOYMENT_ID VARCHAR(255) NOT NULL
 );
 
-CREATE TABLE URI_APP_REGISTRY  (
+CREATE TABLE URI_REGISTRY  (
 	NAME VARCHAR(255) NOT NULL PRIMARY KEY,
 	URI VARCHAR(255) NOT NULL
 );
